feat(payouts): Refactor KYC and improve withdrawal flow   This commit refactors the KYC process and enhances the withdrawal functionality.   Key changes:    - Refactored KYC page to a single-step form    - Implemented file uploads for KYC documents    - Updated withdrawal form to fetch user's events and revenue    - Added useEventRevenue hook to fetch event revenue    - Updated payout service to handle FormData for KYC submissions    - Removed mock payout service and integrated with the actual API
