+++ src/services/event.service.ts
@@ -24,6 +24,7 @@ import type {
 } from "@/types/analytics";
 
 import { apiClient } from "./api.service";
+import type { EventRevenueResponse } from "@/types/event";
 
 /**
  * Event service for handling all event-related operations
@@ -171,6 +172,13 @@ class EventService {
 
   public async cancelEvent(id: string): Promise<EventResponse> {
     return apiClient.post<EventResponse>(`${this.baseUrl}/${id}/cancel`);
+  }
+
+  /**
+   * Event revenue
+   */
+  public async getEventRevenue(id: string): Promise<EventRevenueResponse> {
+    return apiClient.get<EventRevenueResponse>(`${this.baseUrl}/${id}/revenue`);
   }
 
   /**
